+++ auto_tests/set_status_message_test.c
@@ -59,7 +59,7 @@ static void test_set_status_message(void)
         c_sleep(ITERATION_INTERVAL);
     }
 
-    printf("toxes are online, took %ld seconds\n", time(nullptr) - cur_time);
+    printf("toxes are online, took %lu seconds\n", (unsigned long)(time(nullptr) - cur_time));
     const time_t con_time = time(nullptr);
 
     while (tox_friend_get_connection_status(tox1, 0, nullptr) != TOX_CONNECTION_UDP ||
@@ -70,7 +70,7 @@ static void test_set_status_message(void)
         c_sleep(ITERATION_INTERVAL);
     }
 
-    printf("tox clients connected took %ld seconds\n", time(nullptr) - con_time);
+    printf("tox clients connected took %lu seconds\n", (unsigned long)(time(nullptr) - con_time));
 
     TOX_ERR_SET_INFO err_n;
     tox_callback_friend_status_message(tox2, status_callback);
@@ -93,7 +93,7 @@ static void test_set_status_message(void)
     ck_assert_msg(memcmp(cmp_status, STATUS_MESSAGE, sizeof(STATUS_MESSAGE)) == 0,
                   "status message not correct");
 
-    printf("test_set_status_message succeeded, took %ld seconds\n", time(nullptr) - cur_time);
+    printf("test_set_status_message succeeded, took %lu seconds\n", (unsigned long)(time(nullptr) - cur_time));
 
     tox_kill(tox1);
     tox_kill(tox2);
